pyproject.toml
[project]
name = 'file-validation'
version = '0.0.1'
description = 'Python module to validate csv files.'
authors = [
    {name = 'Emiliano Herrera', email = 'eherrera1331@gmail.com'},
]
dependencies = [
    'pyyaml',
    'pandas',
]
